unzip-mem: use the CVE-2018-17828 input checks
Sanitize the input by the remove_dotdotslash function in unzip-mem.c
The unzzip* files should already be safe from the CVE, only the
unzip-mem seems to be vulnerable.

This commit copies the remove_dotdotslah function from the unzzip.c to
unzip-mem.c. While this is not ideal, I suppose that moving the function
to some auxiliary file that would be used by both the unzzip* and
unzip-mem would be too convoluted solution.